move failure report to its own module
